move useAutoScroll to hooks
